chore: move redux directory out of app directory
