+++ src/entrypoints/publishing/redux/reducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 
-import CmsArticleReducer from '../../../../components/with-cms-article/redux/reducer';
+import CmsArticleReducer from '../../../components/with-cms-article/redux/reducer';
 
 export default combineReducers({
   CmsArticleReducer
