+++ src/entrypoints/publishing/redux/saga.ts
@@ -1,6 +1,6 @@
 import { all } from 'redux-saga/effects';
 
-import cmsArticleSaga from '../../../../components/with-cms-article/redux/saga';
+import cmsArticleSaga from '../../../components/with-cms-article/redux/saga';
 
 export default function* saga() {
   yield all([cmsArticleSaga()]);
